requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
Pillow==10.4.0
supabase==2.5.1
